HBASE-22861 [WAL] Merged region should get its WAL according to WALProvider
